+++ RacingLeagueHub.Application/Identity/Authentication/RecoveryCodes/Persistence/IUserRecoveryCodeCommands.cs
@@ -5,5 +5,5 @@ namespace RacingLeagueHub.Application.Identity.Authentication.RecoveryCodes.Pers
 public interface IUserRecoveryCodeCommands
 {
     Task AddRangeAsync(IEnumerable<UserRecoveryCode> userRecoveryCodes, CancellationToken ct = default);
-    Task DeleteForUserAsync(long userId, CancellationToken ct = default);
+    Task DeleteForUserAsync(int userId, CancellationToken ct = default);
 }
