+++ RacingLeagueHub.Application/Identity/Authentication/RefreshTokens/Persistence/IRefreshTokenCommands.cs
@@ -4,7 +4,7 @@ namespace RacingLeagueHub.Application.Identity.Authentication.RefreshTokens.Pers
 
 public interface IRefreshTokenCommands
 {
-    Task<RefreshToken> AddAsync(string refreshToken, long userId, DateTime expiresAt, CancellationToken ct = default);
+    Task<RefreshToken> AddAsync(string refreshToken, int userId, DateTime expiresAt, CancellationToken ct = default);
     Task RevokeRefreshTokenAsync(string refreshToken, CancellationToken ct = default);
     Task SaveChangesAsync(CancellationToken ct = default);
 }
