+++ RacingLeagueHub.Application/Teams/ITeamService.cs
@@ -5,9 +5,9 @@ namespace RacingLeagueHub.Application.Teams;
 
 public interface ITeamService
 {
-    Task<TeamDto?> GetByIdAsync(long id, CancellationToken ct);
+    Task<TeamDto?> GetByIdAsync(int id, CancellationToken ct);
     Task<PagedResult<TeamDto>> GetPagedAsync(int page, CancellationToken ct);
     Task<TeamDto> AddAsync(CreateTeamDto dto, CancellationToken ct);
-    Task<TeamDto?> UpdateAsync(long id, UpdateTeamDto dto, CancellationToken ct);
-    Task<bool> DeleteAsync(long id, CancellationToken ct);
+    Task<TeamDto?> UpdateAsync(int id, UpdateTeamDto dto, CancellationToken ct);
+    Task<bool> DeleteAsync(int id, CancellationToken ct);
 }
