+++ RacingLeagueHub.Application/TrackLayouts/Persistence/ITrackLayoutCommands.cs
@@ -5,5 +5,5 @@ namespace RacingLeagueHub.Application.TrackLayouts.Persistence;
 public interface ITrackLayoutCommands
 {
     Task<TrackLayoutDto> AddAsync(CreateTrackLayoutDto dto, CancellationToken ct);
-    Task<TrackLayoutDto?> UpdateAsync(long id, UpdateTrackLayoutDto dto, CancellationToken ct);
+    Task<TrackLayoutDto?> UpdateAsync(int id, UpdateTrackLayoutDto dto, CancellationToken ct);
 }
