+++ RacingLeagueHub.Application/TrackLayouts/Persistence/ITrackLayoutQueries.cs
@@ -5,6 +5,6 @@ namespace RacingLeagueHub.Application.TrackLayouts.Persistence;
 
 public interface ITrackLayoutQueries
 {
-    Task<TrackLayoutDto?> GetByIdAsync(long id, CancellationToken ct);
+    Task<TrackLayoutDto?> GetByIdAsync(int id, CancellationToken ct);
     Task<PagedResult<TrackLayoutDto>> GetPagedAsync(int page, int pageSize, CancellationToken ct);
 }
