+++ RacingLeagueHub.Application/Tracks/Persistence/ITrackQueries.cs
@@ -6,7 +6,7 @@ namespace RacingLeagueHub.Application.Tracks.Persistence;
 
 public interface ITrackQueries
 {
-    Task<TrackDto?> GetByIdAsync(long id, CancellationToken ct);
+    Task<TrackDto?> GetByIdAsync(int id, CancellationToken ct);
     Task<PagedResult<TrackDto>> GetPagedAsync(int page, int pageSize, CancellationToken ct);
     Task<List<LookupDto>> GetLookupsAsync(CancellationToken ct);
 }
