+++ RacingLeagueHub.Domain/Entities/Seasons/SeasonPoints.cs
@@ -4,7 +4,7 @@ namespace RacingLeagueHub.Domain.Entities.Seasons;
 
 public class SeasonPoints : EntityBase
 {
-    public long SeasonId { get; set; }
+    public int SeasonId { get; set; }
     public SessionType SessionType { get; set; }
     public short Position { get; set; }
     public short Points { get; set; }
