+++ RacingLeagueHub.Infrastructure/Persistence/Seasons/SeasonQueries.cs
@@ -45,7 +45,7 @@ internal class SeasonQueries : ISeasonQueries
     }
 
     public async Task<PagedResult<SeasonDto>> GetLeagueSeasonsAsync(
-        long leagueId,
+        int leagueId,
         int page = 1,
         int pageSize = 10,
         CancellationToken ct = default)
